Uploading the VICTRE calibration phantom
This calibration phantom was created by Chrisitan Graff in 2017. Simulations of this phantom are shown in Figures 6 and 7 of our Computer Physics and Communications paper (2021). The phantom has 2000x2000x900 voxels of size 50 micron (4.5 cm thickness). Each voxel is stored as one byte (unsigned char type in C). It can be easily visualized in imageJ, for example.